requirements.txt
scipy==1.8.1
scikit-learn==1.1.2
numpy==1.22.4
matplotlib==3.1.2
opencv-python==4.6.0.66
tqdm==4.60.0
Pillow==8.2.0
h5py==2.10.0
lmdb==1.3.0
torch==1.2.0
torchvision==0.4.0
# postgresql
psycopg2==2.9.5
SQLAlchemy-1.4.46
# export restful api
flask==2.2.2